Revert "Merge branch 'revert-84194a45' into 'develop'"

See merge request mabarahona/tp-final-project-2022-11-10!30

+++ src/app/movie-details/components/movie-details-rating/movie-details-rating.component.ts
@@ -1,7 +1,9 @@
 import { Component, Input, OnInit } from '@angular/core';
+import { MatDialog, MatDialogRef } from '@angular/material/dialog';
 import { Store } from '@ngrx/store';
 import { Observable, take } from 'rxjs';
 import { AppState } from 'src/app/app.store';
+import { ConfirmationDialogComponent } from 'src/app/shared/components/confirmation-dialog/confirmation-dialog.component';
 import { AddRating, RemoveRating } from '../../store/specific-movie.actions';
 
 @Component({
@@ -14,7 +16,9 @@ export class MovieDetailsRatingComponent {
   rates = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10];
   @Input() logged$!: Observable<boolean>;
   @Input() movieId!: number;
-  constructor(private store: Store<AppState>) {}
+  confirmDialog!: MatDialogRef<ConfirmationDialogComponent>;
+
+  constructor(private store: Store<AppState>, private dialog: MatDialog) {}
 
   addRate(rate: number) {
     this.logged$.pipe(take(1)).subscribe(loggedIn => {
@@ -29,8 +33,21 @@ export class MovieDetailsRatingComponent {
   removeRate() {
     this.logged$.pipe(take(1)).subscribe(loggedIn => {
       if (loggedIn) {
-        this.store.dispatch(RemoveRating({ payload: this.movieId }));
+        this.openConfirmationDialog("Are you sure you want to remove this movie from the watchlist?").subscribe(result => {
+          if(result){
+            this.store.dispatch(RemoveRating({ payload: this.movieId }));
+          }
+        })
       }
     });
+  }
+
+  openConfirmationDialog(question: string): Observable<boolean> {
+    this.confirmDialog = this.dialog.open(ConfirmationDialogComponent, {
+      disableClose: true,
+    });
+    this.confirmDialog.componentInstance.dialogMessage =
+      question;
+    return this.confirmDialog.afterClosed();
   }
 }
